refactor: update auth state
