+++ src/entities/state/index.ts
@@ -1,6 +1,3 @@
-export { State } from "./model/StatePublisher";
-
-// TODO: get rid of this state altogether, and use StatePublisher instead
-export { State as StateAuth } from "./model/State";
+export { State } from "./model/State";
 
 export { type Publisher, type Observer } from "./model/Observer";
